% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectoryPlots.R
\name{clustPlots}
\alias{clustPlots}
\title{Function plotting patient trajectories without OUT OF COHORT periods}
\usage{
clustPlots(patStateLevel, pathClicked, patPaths, colorsDef)
}
\arguments{
\item{patStateLevel}{data frame from trajectoryDataPrep}

\item{pathClicked}{reactive input from clicking on the sunburst}

\item{patPaths}{data frame from trajectoryDataPrep}

\item{colorsDef}{from trajectoryDataPrep}
}
\description{
Function plotting patient trajectories without OUT OF COHORT periods
}
