% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectoryPlots.R
\name{cutStart}
\alias{cutStart}
\title{Function cutting data from selected start}
\usage{
cutStart(patStateLevel, startState)
}
\arguments{
\item{patStateLevel}{data frame from trajectoryDataPrep}

\item{startState}{reactive input from user input}
}
\description{
Function cutting data from selected start
}
