% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{stats3}
\alias{stats3}
\title{Average length or times of 3rd STATE level}
\usage{
stats3(pathClicked, patPaths)
}
\arguments{
\item{pathClicked}{reactive input from clicking on the sunburst}

\item{patPaths}{data frame with all patients and their paths}
}
\description{
Average length or times of 3rd STATE level
}
