% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{stats4}
\alias{stats4}
\title{Average length or times of 4th STATE level}
\usage{
stats4(pathClicked, patPaths)
}
\arguments{
\item{pathClicked}{reactive input from clicking on the sunburst}

\item{patPaths}{data frame with all patients and their paths}
}
\description{
Average length or times of 4th STATE level
}
