% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectoryDataPrep.R
\name{trajectoryDataPrep}
\alias{trajectoryDataPrep}
\title{Function preparing labels, colors and intermediate data frames "patStateLevel", "patPaths" and "freqPaths"}
\usage{
trajectoryDataPrep(inputData)
}
\arguments{
\item{inputData}{data read in when running function "trajectoryViz"}
}
\description{
Function preparing labels, colors and intermediate data frames "patStateLevel", "patPaths" and "freqPaths"
}
