% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory_viz-package.R
\docType{package}
\name{trajectoryViz-package}
\alias{trajectoryViz-package}
\title{Clinical Trajectory Visualisation}
\description{
Visualize Trajectories (drug or state sequences) and extract detailed paths
}
