% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectoryViz.R
\name{trajectoryViz}
\alias{trajectoryViz}
\title{The main function running the Shiny App}
\usage{
trajectoryViz(inputData = NULL)
}
\arguments{
\item{inputData}{THE INPUT DATA as read_csv("dataASCH.csv")}
}
\description{
The main function running the Shiny App
}
